lejur-api/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.36
psycopg2-binary==2.9.9
alembic==1.13.3
python-dotenv==1.0.1
redis==5.0.8
pydantic==2.9.2
email-validator==2.2.0
